requirements.txt
openai>=1.0.0
anthropic>=0.3.0
groq>=0.3.0
google-cloud-aiplatform>=1.0.0
cohere>=4.0.0
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
python-dotenv>=1.0.0
wandb>=0.15.0
tqdm>=4.65.0
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
